unify  gather kernel about host/x86/arm test=develop
